fix(node:process): chdir error path/dest/errno match Node (#2135)
process.chdir() to a missing/invalid directory threw an error whose
fields diverged from Node: `path` was set to the target, and `dest`
and `errno` were absent. Node's libuv chdir error sets `path` to the
*source* (cwd), `dest` to the failed target, and `errno` to the
negative libuv code (ENOENT=-2 etc.).

Set `path` = cwd, and attach `dest` (target) and `errno` (negative
libuv code) via the per-error user-prop side table. The message itself
already matched Node. Backs test-process-chdir-errormessage.js (#2135).

chdir + its error construction moved to a new `os::chdir` submodule so
os.rs stays under the 2000-line file-size gate.
